--- crates/test_applets/Cargo.toml
@@ -1,69 +1,0 @@
-[package]
-edition = "2021"
-name = "we_test_applets"
-version = "0.0.1"
-description = "A CLI to test we applets"
-license-file = "LICENSE_CAL-1.0"
-homepage = "https://developer.holochain.org"
-documentation = "https://docs.rs/holochain_cli_launch"
-repository = "https://github.com/lightningrodlabs/we"
-readme = "README.md"
-categories = ["command-line-utilities", "development-tools"]
-
-[[bin]]
-name = "we-test-applets"
-path = "src/bin/we-test-applets.rs"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[build-dependencies]
-tauri-build = { version = "1.5.0", features = [] }
-
-[dependencies]
-anyhow = "1.0"
-futures = "0.3"
-
-we-alpha = { path = "../../src-tauri" }
-we_types = { workspace = true }
-
-holochain_cli_sandbox = "0.2.3-beta-rc.1"
-holochain_types = "0.2.3-beta-rc.0"
-holochain_trace = "0.2.2"
-holochain_client = { git = "https://github.com/matthme/holochain-client-rust", branch = "holochain-0.2.3-beta-rc.1" }
-fixt = "0.2.2"
-
-clap = { version = "4.0", features = ["derive", "env"] }
-log = "0.4.14"
-mime_guess = "2.0.4"
-notify = "5.0.0"
-observability = "0.1.3"
-portpicker = "0.1.1"
-tempdir = "0.3.7"
-thiserror = "1.0.22"
-
-serde = { version = "1", features = ["derive"] }
-tauri = { version = "=1.5", features = [
-  "clipboard-write-text",
-  "dialog-all",
-  "icon-png",
-  "notification",
-  "process-command-api",
-  "shell-open",
-] }
-tauri-plugin-localhost = "0.1.0"
-
-tokio = "1.21.2"
-tracing = "0.1"
-url = "2.3.1"
-url2 = "0.0.6"
-zip = "0.5.0"
-
-uuid = "1.5.0"
-
-[features]
-# by default Tauri runs in production mode
-# when `tauri dev` runs it is executed with `cargo run --no-default-features` if `devPath` is an URL
-default = ["custom-protocol"]
-# this feature is used used for production builds where `devPath` points to the filesystem
-# DO NOT remove this
-custom-protocol = ["tauri/custom-protocol"]
