--- crates/test_applets/tauri.conf.json
@@ -1,45 +1,0 @@
-{
-  "build": {
-    "devPath": "../../ui/app/dist",
-    "distDir": "../../ui/app/dist"
-  },
-  "tauri": {
-    "allowlist": {
-      "clipboard": {
-        "writeText": true
-      },
-      "dialog": {
-        "all": true
-      },
-      "shell": {
-        "open": true
-      }
-    },
-    "bundle": {
-      "active": true,
-      "icon": [
-        "../../src-tauri/icons/32x32.png",
-        "../../src-tauri/icons/128x128.png",
-        "../../src-tauri/icons/128x128@2x.png",
-        "../../src-tauri/icons/icon.icns",
-        "../../src-tauri/icons/icon.ico"
-      ],
-      "externalBin": [
-        "../../src-tauri/bins/holochain-v0.2.3-beta-rc.1",
-        "../../src-tauri/bins/lair-keystore-v0.3.0"
-      ],
-      "identifier": "org.lightningrodlabs.we-alpha",
-      "resources": [
-        "../../src-tauri/icons/icon_priority_high_32x32.png",
-        "../../src-tauri/icons/icon_priority_medium_32x32.png",
-        "../../src-tauri/icons/32x32.png"
-      ],
-      "targets": "all"
-    },
-    "security": {
-      "csp": null,
-      "dangerousDisableAssetCspModification": true
-    },
-    "windows": []
-  }
-}
