+++ flake.nix
@@ -27,7 +27,22 @@
           , lib
           , self'
           , ...
-          }: {
+          }: 
+          let 
+            libraries = with pkgs;[
+              libayatana-appindicator
+              webkitgtk
+              libappindicator-gtk3
+              gtk3
+              cairo
+              gdk-pixbuf
+              glib
+              dbus
+              openssl_3
+              librsvg
+            ];
+          in
+          {
             devShells.default = pkgs.mkShell {
               inputsFrom = [ inputs'.holochain.devShells.holonix ];
               packages = with pkgs; [
@@ -35,7 +50,6 @@
                 cargo-nextest
                 xvfb-run
               ];
-
               buildInputs = (with pkgs; [
                 openssl
 
@@ -44,10 +58,17 @@
               ])
               ++ (lib.optionals pkgs.stdenv.isLinux
                 (with pkgs; [
+                  curl
+                  wget
+                  dbus
+                  openssl_3
+                  glib
+                  libsoup
+                  libappindicator-gtk3
+                  librsvg
                   webkitgtk.dev
                   gdk-pixbuf
                   gtk3
-                  libayatana-appindicator
                   # Video/Audio data composition framework tools like "gst-inspect", "gst-launch" ...
                   gst_all_1.gstreamer
                   # Common plugins like "filesrc" to combine within e.g. gst-launch
@@ -60,7 +81,6 @@
                   gst_all_1.gst-libav
                   # Support the Video Audio (Hardware) Acceleration API
                   gst_all_1.gst-vaapi
-                  libsoup_3
                 ]))
               ++ lib.optionals pkgs.stdenv.isDarwin
                 (with pkgs; [
@@ -93,6 +113,9 @@
                 export GIO_MODULE_DIR=${pkgs.glib-networking}/lib/gio/modules/
                 export GIO_EXTRA_MODULES=${pkgs.glib-networking}/lib/gio/modules
                 export WEBKIT_DISABLE_COMPOSITING_MODE=1
+                export LD_LIBRARY_PATH=${pkgs.lib.makeLibraryPath libraries}:$LD_LIBRARY_PATH
+                export XDG_DATA_DIRS=${pkgs.gsettings-desktop-schemas}/share/gsettings-schemas/${pkgs.gsettings-desktop-schemas.name}:${pkgs.gtk3}/share/gsettings-schemas/${pkgs.gtk3.name}:$XDG_DATA_DIRS
+
               '';
             };
           };
