+++ package.json
@@ -15,7 +15,7 @@
     "network:local": "concurrently \"hc run-local-services --bootstrap-port 35355 --signal-port 35356\" \"BOOTSTRAP_PORT=35355 SIGNAL_PORT=35356 npm run network\"",
     "network": "npm run build:happ && npm run build -w @lightningrodlabs/we-applet && concurrently \"npm run start:ui\" \"cross-env ADMIN_PORT=10001 npm run start:agent\" \"cross-env DISABLE_DEEP_LINK=true ADMIN_PORT=10002 npm run start:agent\"",
     "start:ui": "concurrently -k \"npm run build:watch -w @lightningrodlabs/we-applet\" \"npm run start -w we-alpha\"",
-    "start:agent": "concurrently \"npm run tauri:dev\" \"npm run publish:applets\"",
+    "start:agent": "concurrently \"npm run publish:applets\" \"npm run tauri:dev\" ",
     "tauri:dev": "cross-env RUST_LOG=info WASM_LOG=warn npm run tauri dev -- -- -- --profile ${ADMIN_PORT}",
     "publish:applets": "run-singleton \"node scripts/publish-applets.js\"",
     "build:libs": "npm run build -w @lightningrodlabs/we-applet -w grapes-editor -w @lightningrodlabs/attachments",
@@ -34,7 +34,7 @@
   "devDependencies": {
     "@holochain-playground/cli": "^0.1.1",
     "@open-wc/eslint-config": "^12.0.0",
-    "@tauri-apps/cli": "^1.5.2",
+    "@tauri-apps/cli": "^1.5.6",
     "@typescript-eslint/parser": "^6.7.5",
     "adm-zip": "^0.5.10",
     "concurrently": "^6.2.1",
