+++ src-tauri/src/config.rs
@@ -1,5 +1,5 @@
 #[derive(Clone)]
 pub struct WeConfig {
-    pub network_seed: Option<String>,
+    pub default_apps_network_seed: Option<String>,
     pub applets_ui_port: u16,
 }
