+++ src-tauri/src/default_apps.rs
@@ -4,8 +4,11 @@ use holochain_client::{AdminWebsocket, InstallAppPayload};
 use holochain_types::web_app::WebAppBundle;
 use tauri::AppHandle;
 
-use crate::{config::WeConfig, error::WeResult, filesystem::{WeFileSystem, breaking_app_version, UiIdentifier}};
-
+use crate::{
+    config::WeConfig,
+    error::WeResult,
+    filesystem::{breaking_app_version, UiIdentifier, WeFileSystem},
+};
 
 pub fn devhub_app_id(app_handle: &AppHandle) -> String {
     format!("DevHub-{}", breaking_app_version(app_handle))
@@ -16,7 +19,7 @@ pub fn appstore_app_id(app_handle: &AppHandle) -> String {
 }
 
 pub fn network_seed(app_handle: &AppHandle, config: &WeConfig) -> String {
-    let network_seed = if let Some(network_seed) = &config.network_seed {
+    let network_seed = if let Some(network_seed) = &config.default_apps_network_seed {
         network_seed.clone()
     } else if cfg!(debug_assertions) {
         format!("lightningrodlabs-we-dev")
@@ -61,7 +64,10 @@ pub async fn install_default_apps_if_necessary(
 
         we_fs
             .ui_store()
-            .extract_and_store_ui(UiIdentifier::Other(appstore_app_id(app_handle)), &appstore_hub_bundle)
+            .extract_and_store_ui(
+                UiIdentifier::Other(appstore_app_id(app_handle)),
+                &appstore_hub_bundle,
+            )
             .await?;
     }
