+++ src-tauri/src/filesystem.rs
@@ -1,4 +1,3 @@
-use std::fmt::Display;
 use std::path::{Path, PathBuf};
 use std::{fs, io::Write};
