+++ src-tauri/tauri.conf.json
@@ -18,7 +18,17 @@
           "multiple": false
         },
         {
-          "name": "network-seed",
+          "name": "default-apps-network-seed",
+          "takesValue": true,
+          "multiple": false
+        },
+        {
+          "name": "test-applets",
+          "takesValue": true,
+          "multiple": true
+        },
+        {
+          "name": "test-applets-network-seed",
           "takesValue": true,
           "multiple": false
         }
@@ -56,7 +66,11 @@
         "bins/lair-keystore-v0.3.0"
       ],
       "identifier": "org.lightningrodlabs.we-alpha",
-      "resources": ["icons/icon_priority_high_32x32.png", "icons/icon_priority_medium_32x32.png", "icons/32x32.png"],
+      "resources": [
+        "icons/icon_priority_high_32x32.png",
+        "icons/icon_priority_medium_32x32.png",
+        "icons/32x32.png"
+      ],
       "targets": "all"
     },
     "security": {
